Routes students to CCI Printers located on UNC Chapel Hill Campus
This HTML page uses the Google Maps JavaScript API. All CCI Printers on
the UNC Campus are mapped with markers. When one of these markers is
clicked, a route leading from the user's current location to the
marker's location appears. An info window will also open to display
details about where each printer can be found within its building and to
which students use of the printer is restricted , if any.